+++ nf_core/lint/schema_description.py
@@ -1,7 +1,5 @@
 #!/usr/bin/env python
 
-from logging import warn
-
 import nf_core.schema